+++ packages/jsActions/nanoflow-actions-web/CHANGELOG.md
@@ -4,3 +4,27 @@ All notable changes to this widget will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.2.0] - Unreleased
+
+### Added
+- Added the possibility to customize the picture quality from the "Take a picture" action.
+
+### Changed
+- Changed the confirmation screen to hide the control buttons after confirmation while the picture is loading in the "Take a picture" action.
+- Changed the "Take a picture" action to hide the action buttons after taking a picture while waiting for the picture to be processed.
+
+## [2.1.1] - 2021-07-01
+
+### Changed
+- We changed the category of 'Take picture' action to 'Web camera'.
+
+## [2.1.0] - 2021-05-28
+
+### Added
+- We've added a JavaScript action to allow take picture in PWA and web apps.
+
+## [2.0.0] - 2021-04-20
+
+### Added
+- Adding compatibility with Studio Pro 9.